chore: address IntelliJ inspection warnings (safe sweep) (#155)
* chore: address IntelliJ inspection warnings (safe sweep)

Mechanical clean-up driven by a multi-agent IDE-inspection survey
(workflow over 12 lib modules, 375 findings total). Restricts itself
to the categories that are safe to fix without judgment.

Three categories addressed:

1. unused-pattern-variable (15 sites): rename `__` to `_` in
   sealed-Result pattern matches. Java 22+ unnamed pattern;
   mechanical with no behavior change.
2. redundant-cast (13 → 12 sites): remove casts the compiler can
   infer from context. The (long) cast on
   DispatcherIntegrationTest.java:105 stays — it's the boxing path
   from int to Map<TopicPartition, Long>, where IntelliJ's
   "redundant" verdict was actually wrong (compile-fail without it).
3. redundant-suppression (14 sites): delete stale @SuppressWarnings
   annotations that no longer match any real warning.

Plus two unresolved-javadoc-symbol fixes (ERROR severity):
- KPipeDlqTest: qualify [Builder#withDeadLetterQueue(...)] →
  [KPipeConsumer.Builder#withDeadLetterQueue(...)]
- ParallelDispatcherTest: qualify [#close()] →
  [ParallelDispatcher#close()]

What's NOT in this PR (deferred to per-site review):

- unused-parameter (51): almost all are deliberate-ignore in test
  lambdas; needs `_` rename per-site, not a blanket fix.
- try-with-resources (39): rewriting integration tests would change
  Handle / KPipeConsumer shutdown semantics.
- busy-wait (17): intentional poll-until-condition loops in shutdown
  drain and KEY_ORDERED retry paths.
- nullable-handling (16): per-site judgment, especially AvroFormat
  registry-mode hot path.
- unused-method (11): public fluent-facade methods kept as escape
  hatches; documented in docs/escape-hatches.md.
- unchecked-assignment (9): Mockito raw-type captures and Kafka
  ConsumerRecord/ProducerRecord generics need per-site decisions.

Test plan:
- ./gradlew test  (green locally)
- ./gradlew spotlessCheck  (green)
- ./gradlew compileJava compileTestJava  (green)

* refactor: dedup common consumer-config block in DefaultSink + DefaultBatchSink

Both sinks ended their start() with the same 6-line "apply optional
consumer config to the builder" block (retry, backpressure, metrics,
errorHandler, deadLetterTopic, pollTimeout). IntelliJ flagged it as a
9-line code-duplication.

Extract to a package-private DefaultStream.applyCommonConsumerConfig
helper. DefaultSink calls it and then adds its tracer +
circuit-breaker (those two are intentionally not on DefaultBatchSink
today — adding them would be a behavior change, deferred).

* refactor: dedup JMH teardown boilerplate + fix javadoc URL lints

Two follow-ups from the IDE-inspection sweep, this time covering
benchmarks/ and examples/ (the first pass only hit lib/).

**Benchmark teardown dedup.** Two invocation-context inner classes in
ParallelProcessingBenchmarkInfrastructure (Raw + Share) had the same
17-line shutdown block: lower the `running` flag, join the poll loop,
shutdown the executor, await termination (return ignored). IntelliJ
flagged it as duplicated-code; the ignored-return was a separate
finding sitting in the same blocks. Extract to
`stopPollLoopAndExecutor(AtomicBoolean, Thread, ExecutorService)` —
single source of truth, and the `awaitTermination` return is now
logged at WARNING (don't silently swallow) so a slow executor doesn't
disappear from the report.

**Plain-text URL in Javadoc.** Two example files had bare
`http://...` URLs inside triple-slash markdown doc comments, which
IntelliJ flags as plain-text-url-in-javadoc. Wrap in angle brackets
so javadoc renders them as autolinks (Markdown URL convention).

**Drop now-redundant Json/Avro exclusion.** PR #153 added a
benchmarks/build.gradle.kts source-set exclusion to unblock JMH
compilation when those benchmarks were broken. PR #154 restored the
benchmarks but #153 merged after, leaving the exclusion in place
against now-compilable files. Remove the dead exclusion block.

* test: pin Handle defaults + DefaultHandle.startAndWrap exception path

Targets the two real coverage gaps in lib/kpipe-api/src/main that
codecov flagged:

1. HandleDefaultsTest — Handle.java was at 0% / 4 tracked lines.
   The defaults aren't decoration: topKeyQueueDepths(n) has a real
   guard (n <= 0 → IllegalArgumentException) and a real return shape
   (immutable empty list), and close() commits to a 5-second
   shutdownGracefully timeout that every try-with-resources caller
   relies on. The test class uses a minimal Handle subclass that
   overrides only the abstract methods so the defaults actually run,
   then pins each: bad n, valid n, default close timeout.

2. DefaultHandleStartAndWrapTest — DefaultHandle.startAndWrap is the
   only non-delegator in the class (everything else just forwards to
   the consumer, which is not worth a mock-against-mock test). The
   factory has a non-trivial error path: if consumer.start() throws,
   it must close the consumer (so AutoCloseable holds when the caller
   never gets a Handle back), and if close() also throws, the
   secondary exception must be addSuppressed onto the original. The
   test pins happy path, start-failure, start+close-failure
   (suppressed exception attached), and null-consumer rejection.

Skipped on purpose:
- DefaultHandle delegators (isHealthy / metrics / awaitShutdown /
  shutdownGracefully) — testing them with mocks verifies the mock,
  not behavior. Coverage % would rise, real safety wouldn't.

+++ lib/kpipe-consumer/src/test/java/io/github/eschizoid/kpipe/consumer/DispatcherIntegrationTest.java
@@ -398,12 +398,12 @@ class DispatcherIntegrationTest {
           () -> "mode " + mode + " left pending offsets on " + tp + ": " + state
         );
         assertEquals(
-          (long) (recordsPerPartition - 1),
+          (recordsPerPartition - 1),
           ((Number) state.get("highestProcessedOffset")).longValue(),
           () -> "mode " + mode + " did not mark all offsets processed on " + tp + ": " + state
         );
         assertEquals(
-          (long) recordsPerPartition,
+          recordsPerPartition,
           ((Number) state.get("nextOffsetToCommit")).longValue(),
           () -> "mode " + mode + " did not advance commit offset on " + tp + ": " + state
         );
